Polished a bit more the motor control code. Some variables were overflowing ints

+++ spinning_lidar_motor_control/teensy_motor_control/ros_comm.h
@@ -14,7 +14,7 @@ using spinning_lidar_motor_control::ChangeTargetVelocity;
 
 
 /************************       Constants       ************************/
-const int BAUD_RATE = 57600;
+const long BAUD_RATE = 57600;
 const char* SENSOR_FRAME = "/laser_axis";
